+++ dist.browser/baseTransaction.js
@@ -31,7 +31,7 @@ var __importDefault = (this && this.__importDefault) || function (mod) {
 };
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.BaseTransaction = void 0;
-var common_1 = __importDefault(require("@ethereumjs/common"));
+var common_1 = __importDefault(require("@sbr/common"));
 var sbr_util_1 = require("sbr-util");
 /**
  * This base class will likely be subject to further