+++ dist.browser/transactionFactory.d.ts
@@ -1,5 +1,5 @@
 /// <reference types="node" />
-import Common from '@ethereumjs/common';
+import Common from '@sbr/common';
 import { default as Transaction } from './legacyTransaction';
 import { default as AccessListEIP2930Transaction } from './eip2930Transaction';
 import { TxOptions, TypedTransaction, TxData, AccessListEIP2930TxData } from './types';