+++ src/types.ts
@@ -1,5 +1,5 @@
 import { BN, AddressLike, BNLike, BufferLike, PrefixedHexString } from 'sbr-util'
-import Common from '@ethereumjs/common'
+import Common from '@sbr/common'
 import { default as Transaction } from './legacyTransaction'
 import { default as AccessListEIP2930Transaction } from './eip2930Transaction'