+++ test/transactionRunner.ts
@@ -1,7 +1,7 @@
 import tape from 'tape'
 import minimist from 'minimist'
 import { toBuffer } from 'sbr-util'
-import Common from '@ethereumjs/common'
+import Common from '@sbr/common'
 import { Transaction } from '../src/'
 import { ForkName, ForkNamesMap, OfficialTransactionTestData } from './types'